requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
psutil==5.9.5
Pillow==10.0.0
moviepy==1.0.3
librosa==0.10.1
soundfile==0.12.1
numpy==1.24.3
scipy==1.11.3